requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
django-allauth
dj-rest-auth
mysqlclient
python-dotenv
requests